Add to BranchingGuidelines.md
Fleshed out the types of branches, and the rules for branching with them.
To facilitate this a Draw.io project, MontanaBranchingStrategyDiagrams.drawio
has been created. A diagram to help explain the branching pattern and rules
was also created. Finally, an image of the branching pattern diagram,
BranchDirection.png was created.